Combine LinodeRow and LinodeRowWithState...
- Extract hasMutationAvailable
- Extract withNotifications
